#43 extracted calendarInitializer from MapInitializer
